Feature decimals

+++ src/state.rs
@@ -1,18 +1,18 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use cosmwasm_std::{CanonicalAddr, Uint128};
+use cosmwasm_std::{CanonicalAddr, Decimal, Uint128};
 use cw_storage_plus::{Item, Map};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct Config {
     pub denom: String,
-    pub bid_margin: u64,
-    pub lota_fee: u64,
+    pub bid_margin: Decimal,
+    pub lota_fee: Decimal,
     pub lota_contract: CanonicalAddr,
-    pub sity_full_rewards: u64,
-    pub sity_partial_rewards: u64,
-    pub sity_fee_registration: u64,
+    pub sity_full_rewards: Decimal,
+    pub sity_partial_rewards: Decimal,
+    pub sity_fee_registration: Decimal,
     pub sity_min_opening: Uint128,
 }
 pub const CONFIG: Item<Config> = Item::new("config");
@@ -28,7 +28,7 @@ pub const STATE: Item<State> = Item::new("state");
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct CharityInfo {
     pub address: CanonicalAddr,
-    pub fee_percentage: u64,
+    pub fee_percentage: Decimal,
 }
 
 /*
